[PHP] Add `PHP.PHP-NTS.8.4` (Non-thread safe builds of PHP 8.4)
Adds support for PHP Non-thread safe builds, which are commonly used
in CLI, and applications that do not need thread-safety.

---

This is a follow-up to the PR that added to `PHP.PHP.8.4`[^1].

Currently, the `PHP.PHP.8.4` package provides thread-safe builds of
PHP, which are mostly suitable for Apache, and other applications that
require a thread-safe build of PHP. Thread-safe builds can be used in
CLI applications as well, but the proposed `PHP.PHP-NTS.8.4` and
non-thread safe builds in general perform faster, and has less resource
impact.

---

There is another PR (#292723) that proposes to add NTS builds as well.

This PR proposes to use the `PHP.PHP-NTS.` namespace, while the linked
PR proposes `PHP.PHP.NTS` namespace.

I believe, the `PHP.PHP-NTS` namespace I propose here is more suitable,
and in-line how PHP packages are named as well.

For example, notice the `-nts` suffix to the version name in NTS builds:

```
 TS: https://windows.php.net/downloads/releases/php-8.4.12-Win32-vs16-x64.zip`
NTS: https://windows.php.net/downloads/releases/php-8.4.12-nts-Win32-vs16-x64.zip`
```

Further, I (@Ayesh) was the author of the initial `PHP.PHP.8.4` package
manifest, and have an automated setup[^2] to add new versions.

[^1]: https://github.com/microsoft/winget-pkgs/pull/196185
[^2]: https://github.com/PHPWatch/php-winget-manifest/actions